feat: product를 구입했을 때 수량 감소 method 추가

+++ server/database/src/main/java/com/twentiethcenturygangsta/database/domain/Product.java
@@ -60,14 +60,14 @@ public class Product {
         this.isMinor = isMinor;
     }
 
-    public void decrease(int quantity) {
+    public void decrease(Long quantity) {
         if (hasQuantity(quantity)) {
             this.quantity = this.quantity - quantity;
         }
         throw new RuntimeException("does not decrease product's quantity");
     }
 
-    public boolean hasQuantity(int quantity) {
+    public boolean hasQuantity(Long quantity) {
         return this.quantity - quantity >= 0;
     }
 }
